locale_facets.tcc (num_get<>::do_get(iter_type, iter_type, ios_base&, ios_base::iostate&, bool&), [...]): -Wconversion and -Woverflow fixes.
2007-01-28  Paolo Carlini  <pcarlini@suse.de>

	* include/bits/locale_facets.tcc (num_get<>::do_get(iter_type,
	iter_type, ios_base&, ios_base::iostate&, bool&),
	num_put<>::_M_insert_int(_OutIter, ios_base&, _CharT, _ValueT)):
	-Wconversion and -Woverflow fixes.
	* include/bits/istream.tcc (operator>>(int&), operator>>(short&)):
	Likewise.
	* include/std/valarray (valarray<>::shift, cshift): Likewise.

From-SVN: r121264

+++ libstdc++-v3/include/bits/istream.tcc
@@ -1,6 +1,7 @@
 // istream classes -*- C++ -*-
 
-// Copyright (C) 1997, 1998, 1999, 2000, 2001, 2002, 2003, 2004, 2005, 2006
+// Copyright (C) 1997, 1998, 1999, 2000, 2001, 2002, 2003, 2004, 2005,
+// 2006, 2007
 // Free Software Foundation, Inc.
 //
 // This file is part of the GNU ISO C++ Library.  This library is free
@@ -121,7 +122,7 @@ _GLIBCXX_BEGIN_NAMESPACE(std)
 	{
 	  if (numeric_limits<short>::min() <= __l
 	      && __l <= numeric_limits<short>::max())
-	    __n = __l;
+	    __n = short(__l);
 	  else
 	    this->setstate(ios_base::failbit);
 	}
@@ -141,7 +142,7 @@ _GLIBCXX_BEGIN_NAMESPACE(std)
 	{
 	  if (numeric_limits<int>::min() <= __l
 	      && __l <= numeric_limits<int>::max())
-	    __n = __l;
+	    __n = int(__l);
 	  else
 	    this->setstate(ios_base::failbit);
 	}
